+++ src/mixins/defaultValues.js
@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 import Mustache from 'mustache';
 import _ from 'lodash';
